requirements.txt
playwright
requests
pillow
pytesseract
python-dotenv
